+++ src/blocks/CTA/__stories__/CTA.stories.tsx
@@ -1,0 +1,41 @@
+import React from 'react';
+
+import {PageConstructor} from '@gravity-ui/page-constructor';
+import {Meta, Story} from '@storybook/react/types-6-0';
+
+import {BlockType, BlogPostData} from '../../../models/common';
+
+import customBlocks from '../../../constructor/blocksMap';
+import {BLOCKS} from '../../../demo/constants';
+import {PostPageContext} from '../../../contexts/PostPageContext';
+import post from '../../../../.mocks/post.json';
+
+import {CTA, CTAProps} from '../CTA';
+
+export default {
+    title: `${BLOCKS}/CTA`,
+    component: CTA,
+    args: {
+        theme: 'light',
+    },
+} as Meta;
+
+type CTAModel = {
+    type: BlockType.CTA;
+} & CTAProps;
+
+const DefaultTemplate: Story<CTAModel> = (args) => (
+    <PostPageContext.Provider value={{post: post as BlogPostData}}>
+        <PageConstructor content={{blocks: [args]}} custom={customBlocks} />
+    </PostPageContext.Provider>
+);
+
+export const Default = DefaultTemplate.bind({});
+
+Default.args = {
+    type: BlockType.CTA,
+    columnCount: 3,
+    paddingBottom: 'l',
+    paddingTop: 'l',
+    items: [],
+};