+++ src/blocks/Media/__stories__/Media.stories.tsx
@@ -1,0 +1,41 @@
+import React from 'react';
+
+import {PageConstructor} from '@gravity-ui/page-constructor';
+import {Meta, Story} from '@storybook/react/types-6-0';
+
+import {BlockType, BlogPostData} from '../../../models/common';
+
+import customBlocks from '../../../constructor/blocksMap';
+import {BLOCKS} from '../../../demo/constants';
+import {PostPageContext} from '../../../contexts/PostPageContext';
+import post from '../../../../.mocks/post.json';
+
+import {Media, MediaFullProps} from '../Media';
+
+export default {
+    title: `${BLOCKS}/Media`,
+    component: Media,
+    args: {
+        theme: 'light',
+    },
+} as Meta;
+
+type MediaModel = {
+    type: BlockType.Media;
+} & MediaFullProps;
+
+const DefaultTemplate: Story<MediaModel> = (args) => (
+    <PostPageContext.Provider value={{post: post as BlogPostData}}>
+        <PageConstructor content={{blocks: [args]}} custom={customBlocks} />
+    </PostPageContext.Provider>
+);
+
+export const Default = DefaultTemplate.bind({});
+
+Default.args = {
+    type: BlockType.Media,
+    paddingBottom: 'l',
+    paddingTop: 'l',
+    text: 'Test',
+    image: 'https://storage.yandexcloud.net/cloud-www-assets/blog-assets/ru/posts/2022/07/cover-digest-june.png',
+};