+++ src/blocks/Suggest/__stories__/Suggest.stories.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+
+import {PageConstructor} from '@gravity-ui/page-constructor';
+import {Meta, Story} from '@storybook/react/types-6-0';
+
+import {BlockType} from '../../../models/common';
+
+import customBlocks from '../../../constructor/blocksMap';
+import {BLOCKS} from '../../../demo/constants';
+import {PostPageContext} from '../../../contexts/PostPageContext';
+import post from '../../../../.mocks/post.json';
+import suggestedPosts from '../../../../.mocks/suggestedPosts.json';
+
+import {Suggest, SuggestFullProps} from '../Suggest';
+
+export default {
+    title: `${BLOCKS}/Suggest`,
+    component: Suggest,
+    args: {
+        theme: 'light',
+    },
+} as Meta;
+
+type SuggestModel = {
+    type: BlockType.Suggest;
+} & SuggestFullProps;
+
+const DefaultTemplate: Story<SuggestModel> = (args) => (
+    <PostPageContext.Provider value={{post, suggestedPosts}}>
+        <PageConstructor content={{blocks: [args]}} custom={customBlocks} />
+    </PostPageContext.Provider>
+);
+
+export const Default = DefaultTemplate.bind({});
+
+Default.args = {
+    type: BlockType.Suggest,
+    paddingBottom: 'l',
+    paddingTop: 'l',
+};