+++ src/index.ts
@@ -5,7 +5,6 @@ export {BlogConstructorProvider} from './constructor/BlogConstructorProvider';
 export {BlogPostPage} from './containers/BlogPostPage/BlogPostPage';
 export {BlogPage} from './containers/BlogPage/BlogPage';
 
-export * from './models/blog';
 export * from './models/common';
 export * from './models/locale';